requirements.txt
python-telegram-bot
instagram-private-api
linkedin-api
threads-api
atproto
pyyaml
requests
loguru
cloudscraper
beautifulsoup4
twikit
aiohttp
python-dotenv
Pillow
pytz
html-to-json
supabase
pydantic
pydantic-settings
